+++ .github/workflows/publish_testpypi.yml
@@ -33,6 +33,11 @@ jobs:
         --sdist
         --wheel
 #        --outdir dist/
+    - name: Store the distribution packages
+      uses: actions/upload-artifact@v4
+      with:
+        name: python-package-distributions
+        path: dist/
     - name: Publish distribution to Test PyPI
       uses: pypa/gh-action-pypi-publish@release/v1
       with:
