+++ tox.ini
@@ -60,8 +60,8 @@ deps =
     ruff
 skip_install = true
 commands =
-    ruff format --check --diff {posargs:super_repo tests}
-    ruff check --diff {posargs:super_repo tests docs}
+    ruff format --check --diff {posargs:src tests}
+    ruff check --diff {posargs:src tests docs}
 
 [tool:pytest]
 # If a pytest section is found in one of the possible config files
